[RISCV][GlobalISel] Legalize division and remainder
Legalize division and remainder. We test for (s7, s8, s16, s32, s48, s64) on rv64 and (s8, s15, s16, s32, s64, s72, s128) on rv64, with and without the +m, +zmmul extensions. We do not handle types with size > 2 x XLen -- these ought to be handled in the IR pass.

Reviewed By: craig.topper

Differential Revision: https://reviews.llvm.org/D157422