[MLIR][NVGPU] Introduce Warpgroup Matrix Descriptor Type
The Warpgroup Matrix Descriptor is a 64-bit integer that holds information about a matrix used by the wgmma instruction.

In this work, a new type is introduced for the descriptor. This enhances the readability of the IR and allows for easier verification using MLIR verification tools.

The type contains a 'memref' related to the descriptor, which is crucial for preserving and conveying information.

Depends on D157382

Reviewed By: qcolombet

Differential Revision: https://reviews.llvm.org/D158403